Allow instantiation without chianti ions
When the `chianti_ions` parameter is not specified,
the `self.chianti_ions` attribute is an empty list and
`chianti_ions_table` is an empty table.
The CASE statement is taken out of the levels query,
and the WHEN condition for chianti ions is added only if
`self.chianti_ions` evaluates to True.