Add fully_ionized_levels to levels not levels_prepared
Create artificial levels for fully ionized ions in `create_levels_lines`.
Previously they were added to the prepared dataframe in the end.
It important to add them to the levels dataframe because they also should be
counted in macro atom references.

As a consequence, `level_id` cannot be used as the index of `levels` because
artificial levels don't have the ids from the database. Although `line_id`
still can be used as an index it is resetted together with `level_id`.